Refactor entire code
